On 9/11/2016 3:30 PM, clinton bowen wrote: > A few things: > 1) I use one term of the Maclaurin series in inequality 1.5 to get > 5/s^\kappa.  I run into a problem when I try to explain the horizontal > displacement by obstacle hexagon rotation.  my result in (1.5) cancels > in inequality (1.10) in the attached paper.  How do I correct this?
That's a good question! After establishing a bound on the vertical
displacement, we can go back to (1.5) and tighten the analysis:
When we already know that the vertical displacement is at most 2 delta_v
(which is at most 300/s^{2kappa-1}), we can plug this value into the
numerator in the first line of (1.5), and improve the bound on gamma_i
to constant/s^{3kappa-1).
…

> 2) Since incorporating the rotation on the central flag, I'm having a
> hard time understanding the vertical argument. It seems lost to me and
> unclear as to why the bound I have has any reasonable justification
> for it to be a valid argument.

I don't see what changes in the argument when the central flag is
allowed to rotate.
The key here is to compare an upper and a lower bound for the vertical
cross-section:
The upper bound is the same as in canonical position, and the lower
bound is independent
of the rotation of the flags (if a central flag rotates, it would only
increase the lower bound).